requirements.txt
#Core packages (prerequisites of the packages below)
pytorch-lightning
torchmetrics
keyboard
torch
pytorch-lightning
torchmetrics
torchaudio

#VAD (Voice Activity Detection) - cobra 
# https://github.com/Picovoice/cobra 
pvcobra
pvrecorder
pyaudio
#ASR (automatic speech recognition) - NEMO
# https://github.com/NVIDIA/NeMo
Cython
hydra
nemo 
nemo_toolkit

ffmpeg-python



#LLM (large language model) - undecided
ChatGPT
openai
python-dotenv

#TTS (text to speech) - GTTS
# https://github.com/coqui-ai/TTS 
gTTS

# additional packages
thread6